Move all polygon related code to polygons
